LPS-85088 Expand UserServiceTest into individual test classes
